+++ nym-vpn-core/crates/nym-vpn-lib/src/lib.rs
@@ -1,6 +1,7 @@
 // Copyright 2023 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: GPL-3.0-only
 
+pub mod paths;
 pub mod storage;
 
 mod adblocker;